+++ homelab_operator/webui/models.py
@@ -4,6 +4,8 @@ import requests
 import socket
 from django.db import models
 from django.utils import timezone
+from django.contrib.contenttypes.fields import GenericForeignKey
+from django.contrib.contenttypes.models import ContentType
 
 from .helpers.system import is_process_running
 
@@ -16,12 +18,23 @@ class UserProfile(models.Model):
                                               help_text='The last selected homelab for this user')
     show_wiki = models.BooleanField(default=True, help_text='Show or hide the wiki in the UI')
     show_networks = models.BooleanField(default=True, help_text='Show or hide networks in the UI')
-    show_ingress = models.BooleanField(default=True, help_text='Show or hide ingress section in the UI')
+    show_ingress = models.BooleanField(default=True,
+                                       help_text='Show or hide ingress section in the UI')
     dark_mode = models.BooleanField(default=False,
                                     help_text='Enable Dark Mode (experimental)')
+    # notifications = ...  # TODO: Implement notifications
 
     def __str__(self):
         return f"Profile of {self.user.username}"
+
+    def get_notifications(self):
+        '''Returns notifications for the user profile.'''
+        maintenance_notifications = self.maintenance_schedules.filter(
+            scheduled_date__gte=timezone.now() - timezone.timedelta(days=2),
+            scheduled_date__lte=timezone.now() + timezone.timedelta(days=7),
+        )
+
+        return [{'title': 'Upcoming Maintenance Plan.', 'date': s.scheduled_date , 'content': s.title,} for s in maintenance_notifications.all()]
 
 class Server(models.Model):
     '''Model representing a server.'''
@@ -32,12 +45,13 @@ class Server(models.Model):
                                help_text='Will be used to check if the server is online')
     mac_address = models.CharField(max_length=17, null=True, blank=True)
     note = models.TextField(null=True, blank=True)
-    network = models.ForeignKey('Network', on_delete=models.CASCADE, null=True,
+    network = models.ForeignKey('Network', on_delete=models.SET_NULL, null=True,
                                 blank=True, related_name='servers')
     user = models.ForeignKey('auth.User', on_delete=models.CASCADE, null=True, blank=True)
     homelab = models.ForeignKey('Homelab', on_delete=models.CASCADE, null=True, blank=True,
                                 related_name='servers')
-    auto_wake = models.BooleanField(default=False, help_text='Automatically wake the server on access')
+    auto_wake = models.BooleanField(default=False,
+                                    help_text='Automatically wake the server on access')
     # TODO Shutdown URL configuration (related_name): implement as shutdown_adapther, which can be
     # a ShutdownURLConfiguration, ShutdownSSLConfiguration, similar ...
 
@@ -205,8 +219,9 @@ class Wiki(models.Model):
     show_services = models.BooleanField(default=True, help_text='Show services in the wiki')
     homelab = models.ForeignKey('Homelab', on_delete=models.CASCADE,
                                 related_name='wiki')  # Homelab expected to only have one wiki
+    pinned_servers = models.ManyToManyField(Server, blank=True,
+                                            help_text='Servers pinned to the wiki for quick access')
     pinned_services = models.ManyToManyField(Service, blank=True,
-                                             related_name='pinned_wikis',
                                              help_text='Services pinned to the wiki for quick access')
 
     def __str__(self):
@@ -394,3 +409,71 @@ class Ingress(models.Model):
                 url = f'http://{url}'
 
             return f'{url}:{service.port}/'
+
+class MaintenancePlan(models.Model):
+    '''Model representing a maintenance plan for servers or services.'''
+    PRIORITY_CHOICES = [
+        ('LOW', 'Low'),
+        ('MEDIUM', 'Medium'),
+        ('HIGH', 'High'),
+        ('CRITICAL', 'Critical'),
+    ]
+    title = models.CharField(max_length=100, help_text='Title of the maintenance plan')
+    description = models.TextField(null=True, blank=True,
+                                   help_text='Description of the maintenance plan')
+    assignee = models.ForeignKey(UserProfile, on_delete=models.CASCADE,
+                                 related_name='maintenance_schedules')
+
+    content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
+    object_id = models.PositiveIntegerField()
+    instance = GenericForeignKey('content_type', 'object_id')
+
+    scheduled_date = models.DateField(help_text='When the maintenance is scheduled')
+    repeat_interval = models.IntegerField(default=0,
+                                          help_text='Repeat interval in days (0 for no repeat)')
+
+    priority = models.CharField(max_length=10, choices=PRIORITY_CHOICES, default='MEDIUM')
+
+    homelab = models.ForeignKey('Homelab', on_delete=models.CASCADE,
+                                related_name='maintenance_plans')
+
+    def __str__(self):
+        return f"{self.title}"
+
+    class Meta:
+        ordering = ['-scheduled_date']
+
+class MaintenanceReport(models.Model):
+    '''Model representing a report for a maintenance task.'''
+    RESULT_CHOICES = [
+        ('OK', 'OK'),
+        ('NOT_OK', 'Not OK'),
+        ('REMARK', 'Remark'),
+        ('ERROR', 'Error'),
+        ('BLOCKED', 'Blocked'),
+        ('SKIPPED', 'Skipped'),
+        ('CANCELLED', 'Cancelled'),
+        ('UNKNOWN', 'Unknown'),
+    ]
+    certifier = models.ForeignKey(
+        UserProfile,
+        on_delete=models.CASCADE,
+        related_name='maintenance_reports',
+        help_text='User who certified the maintenance task'
+    )
+    result = models.CharField(
+        max_length=15,
+        choices=RESULT_CHOICES,
+        default='OK',
+        help_text='Result of the maintenance task'
+    )
+    maintenance_plan = models.ForeignKey(
+        MaintenancePlan,
+        on_delete=models.SET_NULL,
+        related_name='reports',
+        null=True, blank=True,
+    )
+    date = models.DateTimeField(auto_now_add=True,
+                                help_text='Date of the maintenance task')
+    notes = models.TextField(null=True, blank=True,
+                             help_text='Additional notes about the maintenance task')