+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 ## [1.6.0]
 Added:
 - Ingress handling
+- Maintenance functionality
+- Add pinned servers to dashboard
 
 Fixed:
 - Fixed showing unrelated services for pinned services in wiki edit/create
 
 Changed:
 - Mobile Optimizations
+- Miscellaneous improvements
 
 ## [1.5.1]
 Changed: